+++ packages/skeleton-react/src/components/toggle-group/modules/provider.ts
@@ -1,0 +1,12 @@
+import { normalizeProps, useMachine, type PropTypes } from '@zag-js/react';
+import { connect, machine } from '@zag-js/toggle-group';
+import type { Api, Props } from '@zag-js/toggle-group';
+import { useId } from 'react';
+
+export function useToggleGroup(props: Omit<Props, 'id'> = {}): Api<PropTypes> {
+	const service = useMachine(machine, {
+		id: useId(),
+		...props,
+	});
+	return connect(service, normalizeProps);
+}