+++ packages/skeleton-react/test/components/toggle-group/index.test.tsx
@@ -1,0 +1,19 @@
+import Test from './test';
+import { render, screen } from '@testing-library/react';
+import { describe, expect, it } from 'vitest';
+
+describe('ToggleGroup', () => {
+	describe('Root', () => {
+		it('renders', () => {
+			render(<Test />);
+			expect(screen.getByTestId('root')).toBeInTheDocument();
+		});
+	});
+
+	describe('Item', () => {
+		it('renders', () => {
+			render(<Test />);
+			expect(screen.getByTestId('item')).toBeInTheDocument();
+		});
+	});
+});