+++ packages/skeleton-react/test/components/toggle-group/test.tsx
@@ -1,0 +1,9 @@
+import { ToggleGroup } from '@/index';
+
+export default function Test() {
+	return (
+		<ToggleGroup data-testid="root">
+			<ToggleGroup.Item value="item" data-testid="item" />
+		</ToggleGroup>
+	);
+}