+++ packages/skeleton-svelte/src/components/segmented-control/anatomy/control.svelte
@@ -26,5 +26,7 @@
 {#if element}
 	{@render element(attributes)}
 {:else}
-	<div {...attributes}>{@render children?.()}</div>
+	<div {...attributes}>
+		{@render children?.()}
+	</div>
 {/if}