+++ packages/skeleton-svelte/src/components/toggle-group/modules/root-context.ts
@@ -1,0 +1,4 @@
+import type { useToggleGroup } from './provider.svelte';
+import { createContext } from '@/internal/create-context';
+
+export const RootContext = createContext<ReturnType<typeof useToggleGroup>>();