+++ packages/skeleton-svelte/test/components/toggle-group/index.test.ts
@@ -1,0 +1,19 @@
+import Test from './test.svelte';
+import { render, screen } from '@testing-library/svelte';
+import { describe, expect, it } from 'vitest';
+
+describe('ToggleGroup', () => {
+	describe('Root', () => {
+		it('renders', () => {
+			render(Test);
+			expect(screen.getByTestId('root')).toBeInTheDocument();
+		});
+	});
+
+	describe('Item', () => {
+		it('renders', () => {
+			render(Test);
+			expect(screen.getByTestId('item')).toBeInTheDocument();
+		});
+	});
+});